// frt/src/lib/types.d.ts
export interface User {
	userid: string;
	account: string;
	username: string;
	group: string;
	eid: string;
	nickname: string;
	avatar: string;
	signature?: string;
	bio: string;
	notify: string;
	ps: number;
	sessionToken: string;
	password?: string;
	extra?: {
		input_search?: string;
		filter_status?: string;
	};
	perms: unknown;
}

export interface MtcSessionType {
	avatarChangedFlag: number;
	signatureChangedFlag: number;
	tplIds: string[];
	tplIdsForSearch_for_wf: string[];
	delegators: string[];
	tplIdsForSearch_for_todo: string[];
	pickedTodoId: string;
	comment_wfid?: string;
	comments?: string[];
	showpostponed: boolean;
	wfid: string;
	user: User | null;
	org: any;
	orgname: string;
	locale: string;
}
// ...
export interface Workflow {
	_id: string;
	pbo: [string];
	pboat: string;
	endpoint: string;
	endpointmode: string;
	wfid: string;
	tenant: string;
	wftitle: string;
	teamid: string;
	tplid: string;
	status: string;
	statusLabel: string;
	starter: string;
	starterCN: string;
	doc: string;
	createdAt: string;
	updatedAt: string;
	kvarsArr?: kvarDef[];
	attachments: any[];
	rehearsal: boolean;
	pnodeid?: string;
	pworkid?: string;
	cselector?: string[];
	kvars?: any;
	history?: WorkflowHistoryEntry[];
	commentCount: number;
	allowdiscuss: boolean;
	checked: boolean;
}
export type NodeBriefType = {
	nodeid: string;
	title: string;
};

export interface Work {
	_id: string;
	tenant: string;
	doer: string;
	doerCN: string;
	doer_string: string;
	role?: string;
	tplid: string;
	wfid: string;
	wftitle: string;
	todoid: string;
	nodeid: string;
	workid: string;
	cellInfo: string;
	title: string;
	status: string;
	kvars: unknown;
	kvarsArr: ArrayLike<kvarDef>;
	wfstatus: string;
	createdAt: string;
	updatedAt: string;
	instruct: string;
	rehearsal?: boolean;
	wfstarter?: string;
	wfstarterCN?: string;
	wf: Workflow;
	revocable?: boolean;
	returnable?: boolean;
	options?: string[];
	doneby?: string;
	doneat?: date;
	comment?: string;
	comments?: any;
	routingOptions: string[];
	freejump_nodes: NodeBriefType[];
	from_nodeid?: string;
	withsb?: boolean;
	withrvk?: boolean;
	withadhoc?: boolean;
	withcmt?: boolean;
	allowdiscuss: boolean;
	version: string;
	transferable?: boolean;
	mustsign: boolean;
	ai: string;
}

export interface TmapEntry {
	eid: string;
	cn: string;
}

export interface Tmap {
	[k: string]: TmapEntry[];
}

export interface Team {
	_id: string;
	tenant: string;
	teamid: string;
	author: string;
	createdAt: string;
	updatedAt: string;
	tmap: Tmap;
}

export interface kvarDef {
	name: string;
	value: string | number | string[];
	display?: string | number | string[];
	type: InputType;
	label: string;
	placeholder?: string;
	breakrow: boolean;
	id: string;
	options?: string;
	required?: boolean;
	when?: string;
	cssClass?: string;
	hide?: string;
	ui?: string | string[];
	formula?: string;
	wrong_input?: string;
}
// ...
export interface MtcConfirmType {
	title: string;
	body: string;
	buttons: string[];
	callbacks: any[];
}
// ...
export interface UiSectionType {
	tag: boolean;
	search: boolean;
	sort: boolean;
	layout: boolean;
}